moving function get_dataset_key into io module
