+++ lstchain/io/io.py
@@ -1,10 +1,12 @@
 from ctapipe.io import HDF5TableReader
 from ctapipe.io.containers import MCHeaderContainer
 from tables import open_file
+import h5py
 import numpy as np
 
 __all__ = ['read_simu_info_hdf5',
            'read_simu_info_merged_hdf5',
+           'get_dataset_keys',
            ]
 
 def read_simu_info_hdf5(filename):
@@ -53,3 +55,25 @@ def read_simu_info_merged_hdf5(filename):
     combined_mcheader['num_showers'] = num_showers
     return combined_mcheader
 
+
+def get_dataset_keys(filename):
+    """
+    Return a list of all dataset keys in a HDF5 file
+
+    Parameters
+    ----------
+    filename: str - path to the HDF5 file
+
+    Returns
+    -------
+    list of keys
+    """
+    dataset_keys = []
+    def walk(name, obj):
+        if type(obj) == h5py._hl.dataset.Dataset:
+            dataset_keys.append(name)
+
+    with h5py.File(filename,'r') as file:
+        file.visititems(walk)
+
+    return dataset_keys
