+++ scripts/merge_hdf5_files.py
@@ -1,7 +1,7 @@
 import argparse
 import os
-import h5py
 import tables
+from lstchain.io import get_dataset_keys
 
 parser = argparse.ArgumentParser(description="Merge all HDF5 files resulting from parallel reconstructions \
  present in a directory. Every dataset in the files must be readable with pandas.")
@@ -18,29 +18,6 @@ parser.add_argument('--outfile', '-o', action='store', type=str,
                     default='merge.h5')
 
 args = parser.parse_args()
-
-
-def get_dataset_keys(filename):
-    """
-    Return a list of all dataset keys in a HDF5 file
-
-    Parameters
-    ----------
-    filename: str - path to the HDF5 file
-
-    Returns
-    -------
-    list of keys
-    """
-    dataset_keys = []
-    def walk(name, obj):
-        if type(obj) == h5py._hl.dataset.Dataset:
-            dataset_keys.append(name)
-
-    with h5py.File(filename,'r') as file:
-        file.visititems(walk)
-
-    return dataset_keys
